Starting in and around Sea
Starting on a Sea terrain? I don't think so..... Not an island is
allowed....fixed starting on the sea and beside more that two.

Added equals and hashcodes to Coord

Some stuff on InfoPanel....     Not sure if combat against forts will
work as is, but I couldn't get around to testing that