GameMenu update
Changed so that clicking online play prompts the user for a name.
Currently it does not do the following:

- check if a name was entered
- check if the name is already taken
- include a password
- send the name to the right place???

I didn't know where the name should be sent. Check line 253 in GameMenu.
That is where it takes the value of the text in the name textField and
sends it to GameLoop (Added the localPlayer variable in GameLoop for the
time being). You can go ahead and change that to what you need...

Similarly, the local game menu will use a like setup, but with four
names.