Combat done. Class diagram1
Combat is done...as in, the winner gets the terrain, and rolls to see if
fort is kept or downgraded. Exploration shouldn't be long now.
Class Disagram:
- Omitting alot of the classes for simplicity (All the magicEvents,
randomEvents, eventHandlers etc). Do we have to show every class? Or is
there a way to show, "hey, there are more classes here, but there are
alot and we are lazy/don't want to show them all on this one diagram"?

- I removed all but what I thought were the most important relation.
Good/bad idea?